Use an interface to dontaudit file leaks
Doing this in a module other than the one the domain is defined in is a
little unorthodox, but since we can't modify the stock policy this is
the best we can do.

I think this manages to compile without having the relevant bits listed
in gen_requires because the *_domtrans interfaces include them.